Refactor: Bootstrap file
