--- lib/config/bloc/bloc.dart
@@ -1,1 +1,0 @@
-export 'app_bloc_observer.dart';
