+++ lib/config/config.dart
@@ -1,1 +1,2 @@
 export 'constants.dart';
+export 'routes/app_routes.dart';
