+++ lib/main.dart
@@ -1,22 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_downloader/flutter_downloader.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 
 import 'application/app_widget.dart';
-import 'config/bloc/bloc.dart';
+import 'bootstrap.dart';
 import 'core/core.dart';
 import 'core/utils/tool_methods.dart';
-
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await _initialized();
-  BlocOverrides.runZoned(
-    () => runApp(const AppWidget()),
-    blocObserver: AppBlocObserver(),
-  );
+  await bootstrap(() => const Application());
 }
 
 Future<void> _initialized() async {
